+++ Planteskole.WPF/Commands/UpdateCurrentViewModelCommand.cs
@@ -41,9 +41,6 @@ namespace Planteskole.WPF.Commands
                     case ViewType.Database:
                         _navigator.CurrentViewModel = new DatabaseViewModel();
                         break;
-                    case ViewType.Testing:
-                        _navigator.CurrentViewModel = new OrdersViewModel();
-                        break;
                     default:
                         break;
                 }
