+++ Planteskole.WPF/Controls/NavigationBar.xaml
@@ -44,10 +44,6 @@
                          IsChecked="{Binding CurrentViewModel, Mode=OneWay, Converter={StaticResource EqualValueToParameterConverter}, ConverterParameter={x:Type vm:DatabaseViewModel}}"
                          Command="{Binding UpdateCurrentViewModelCommand}" 
                          CommandParameter="{x:Static nav:ViewType.Database}"/>
-            <RadioButton Grid.Column="3" Content="Orders"
-                         IsChecked="{Binding CurrentViewModel, Mode=OneWay, Converter={StaticResource EqualValueToParameterConverter}, ConverterParameter={x:Type vm:OrdersViewModel}}"
-                         Command="{Binding UpdateCurrentViewModelCommand}" 
-                         CommandParameter="{x:Static nav:ViewType.Testing}"/>
 
         </Grid>
     </Grid>
