--- Planteskole.WPF/Controls/StorageExpandableList.xaml
@@ -1,30 +1,0 @@
-﻿<UserControl x:Class="Planteskole.WPF.Controls.StorageExpandableList"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Planteskole.WPF.Controls"
-             xmlns:viewmodels="clr-namespace:Planteskole.WPF.ViewModels"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <UserControl.Resources>
-        <CollectionViewSource x:Key="PlantViewSource"/>
-    </UserControl.Resources>
-    <Grid Loaded="Grid_Loaded" Margin="10">
-        <TreeView Name="trvMenu">
-            <TreeView.Resources>
-                <HierarchicalDataTemplate DataType="{x:Type local:AbstractO}" ItemsSource="{Binding Items}">
-                    <StackPanel Orientation="Horizontal">
-                        <TextBlock Text="{Binding Title}" />
-                    </StackPanel>
-                </HierarchicalDataTemplate>
-                <DataTemplate DataType="{x:Type local:LocationO}">
-                    <TreeViewItem Header="{Binding Title, Mode=TwoWay}">
-                        <DataGrid VerticalScrollBarVisibility="Visible" AlternatingRowBackground="LightGray" AlternationCount="2" ItemsSource="{Binding Source={StaticResource PlantViewSource}}"/>
-                    </TreeViewItem>
-                </DataTemplate>
-            </TreeView.Resources>
-        </TreeView>
-        <Button Grid.Row="1" Content="Save" HorizontalAlignment="Center" Margin="0 10" Click="Button_Click1" Height="20" Width="124"/>
-    </Grid>
-</UserControl>
