--- Planteskole.WPF/ViewModels/StorageExpandableListViewModel.cs
@@ -1,27 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using Planteskole.Domain.Models;
-using Planteskole.WPF.Temporary;
-
-namespace Planteskole.WPF.ViewModels
-{
-    class StorageExpandableListViewModel
-    {
-        public StorageExpandableListViewModel()
-        {
-        }
-    }
-}
