+++ Planteskole.WPF/Views/HomeView.xaml.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Planteskole.Domain.Models;
-using Planteskole.WPF.Temporary;
+﻿using Planteskole.WPF.Temporary;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,60 +17,20 @@ using Planteskole.WPF.ViewModels;
 
 namespace Planteskole.WPF.Views
 {
-    /// <summary>
-    /// Interaction logic for HomeView.xaml
-    /// </summary>
+  
     public partial class HomeView : UserControl
     {
-        private readonly PlantContext _context = new PlantContext();
+        
 
-        private CollectionViewSource PlantViewSource;
-        private CollectionViewSource AreaViewSource;
-        private CollectionViewSource LocationViewSource;
-
+        
         public HomeView()
         {
             InitializeComponent();
+            //PlantViewSource = (CollectionViewSource)FindResource(nameof(PlantViewSource));
+            DataContext = new ViewModels.HomeViewModel();
 
-            PlantViewSource = (CollectionViewSource)FindResource(nameof(PlantViewSource));
-            AreaViewSource = (CollectionViewSource)FindResource(nameof(AreaViewSource));
-            LocationViewSource = (CollectionViewSource)FindResource(nameof(LocationViewSource));
+
         }
-
-        private void Grid_Loaded(object sender, RoutedEventArgs e)
-        {
-            //_context.Database.EnsureDeleted();
-
-            _context.Database.EnsureCreated();
-
-            _context.Plants.Load();
-            _context.Areas.Load();
-            _context.Locations.Load();
-
-            PlantViewSource.Source = _context.Plants.Local.ToObservableCollection();
-            AreaViewSource.Source = _context.Areas.Local.ToObservableCollection();
-            LocationViewSource.Source = _context.Locations.Local.ToObservableCollection();
-
-            this.StorageList.DataContext = new StorageExpandableListViewModel();
-        }
-
-        /*private void Button_Click(object sender, RoutedEventArgs e)
-        {
-            _context.SaveChanges();
-
-            //PlantDataGrid.Items.Refresh();
-        }*/
-
-        /*private void CollectionViewSource_Filter(object sender, FilterEventArgs e)
-        {
-            if ((e.Item as DomainObject).Id == 1)
-            {
-                e.Accepted = true;
-            }
-            else
-            {
-                e.Accepted = false;
-            }
-        }*/
+        
     }
 }
