--- Planteskole.WPF/Views/TestingView.xaml
@@ -1,86 +1,0 @@
-﻿<UserControl x:Class="Planteskole.WPF.Views.TestingView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Planteskole.WPF.ViewModels"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-
-
-
-    <UserControl.Resources>
-        <local:GroupsToTotalConverter x:Key="groupsConverter" />
-    </UserControl.Resources>
-    <Grid>
-        <StackPanel>
-            <DataGrid CanUserAddRows="False" Height="250" ItemsSource="{Binding OrdersView}" >
-                <DataGrid.GroupStyle>
-                    <GroupStyle>
-                        <GroupStyle.HeaderTemplate>
-                            <DataTemplate>
-                                <DockPanel Background="LightBlue">
-                                  
-                                </DockPanel>
-                            </DataTemplate>
-                        </GroupStyle.HeaderTemplate>
-                        <GroupStyle.ContainerStyle>
-                            <Style TargetType="{x:Type GroupItem}" >
-                                <Setter Property="Margin" Value="0,0,0,5" />
-                                <Setter Property="Template">
-                                    <Setter.Value>
-                                        <ControlTemplate TargetType="{x:Type GroupItem}">
-                                            <Expander
-												BorderBrush="#FF002255"
-												IsExpanded="True"
-												Background="LightGray"
-												BorderThickness="0,0,0,1">
-                                                <Expander.Header>
-                                                    <StackPanel Orientation="Horizontal">
-                                                        <TextBlock VerticalAlignment="Center"
-															FontWeight="Bold"
-															Text="{Binding Path=Name}"
-															Width="100" />
-                                                        <TextBlock VerticalAlignment="Center"
-															Text="Item count:" />
-                                                        <TextBlock FontWeight="Bold"
-															Text="{Binding ItemCount}" />
-                                                        <TextBlock VerticalAlignment="Center"
-															TextAlignment="Right"
-															Width="100"
-															Text="Amount" />
-                                                    </StackPanel>
-                                                </Expander.Header>
-                                                <Expander.Content>
-                                                    <ItemsPresenter />
-                                                </Expander.Content>
-                                            </Expander>
-                                        </ControlTemplate>
-                                    </Setter.Value>
-                                </Setter>
-                            </Style>
-                        </GroupStyle.ContainerStyle>
-                    </GroupStyle>
-                    <GroupStyle>
-                        <GroupStyle.HeaderTemplate>
-                            <DataTemplate>
-                                <DockPanel Background="LightBlue">
-                                    <TextBlock Text="{Binding Path=Name}" Foreground="Blue" Margin="30,0,0,0"/>
-                                </DockPanel>
-                            </DataTemplate>
-                        </GroupStyle.HeaderTemplate>
-                    </GroupStyle>
-                </DataGrid.GroupStyle>
-            </DataGrid>
-
-            <StackPanel Orientation="Horizontal" Height="25">
-                <Button Command="{Binding removeGroupCommand}">Remove Group</Button>
-                <Button Command="{Binding groupByLocationCommand}">Location</Button>
-                <Button Command="{Binding groupByAreaCommand}">Area</Button>
-                <Button Command="{Binding saveButtonCommand}">Save</Button>
-                <Button Command="{Binding deleteButtonCommand}">Delete</Button>
-            </StackPanel>
-        </StackPanel>
-        
-    </Grid>
-</UserControl>
